+++ StudyGaemi/SceneDelegate.swift
@@ -19,62 +19,16 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
 
-        let launchScreenVC = LaunchScreenViewController()
+        let launchScreenView = LaunchScreenView()
 
         // Create a new UIWindow
         let window = UIWindow(windowScene: windowScene)
-        window.rootViewController = launchScreenVC
+        window.rootViewController = launchScreenView
         window.makeKeyAndVisible()
         self.window = window
 
         // 로그인 상태 확인 및 적절한 ViewController 설정
         checkLogin()
-        
-        func checkLogin() {
-            print("로그인 유저인지 확인")
-            if let user = Auth.auth().currentUser {
-                print("currentUser임")
-                AuthenticationManager.shared.checkEmailVerifiedForLogin { isEmailVerified in
-                    if isEmailVerified {
-                        self.isLogined = true
-                        print("애플 또는 이메일 자동 로그인 성공: \(user.email ?? "")")
-                        navigateToMainScreen()
-                    } else {
-                        print("자동 로그인 실패. else 시작")
-                        FirestoreManager.shared.getLoginMethod { loginMethod in
-                            if loginMethod == "kakao" {
-                                print("카카오 계정임")
-                                self.isLogined = true
-                                navigateToMainScreen()
-                            } else {
-                                print("카카오 계정 이외의 알 수 없는 무언가")
-                                navigateToLoginScreen()
-                            }
-                        }
-                    }
-                }
-            } else {
-                print("최종 else 진입")
-                navigateToLoginScreen()
-            }
-        }
-        
-        
-        func navigateToMainScreen() {
-            let bottomTabBarVC = BottomTabBarViewController()
-            if let window = self.window {
-                window.rootViewController = bottomTabBarVC
-                window.makeKeyAndVisible()
-            }
-        }
-        
-        func navigateToLoginScreen() {
-            let bottomTabBarVC = LoginViewController()
-            if let window = self.window {
-                window.rootViewController = bottomTabBarVC
-                window.makeKeyAndVisible()
-            }
-        }
     }
 
 
@@ -122,52 +76,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 }
 
-class LaunchScreenViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = UIColor(named: "launchScreenColor")
-        
-        // Logo ImageView
-        let logo = UIImageView(image: UIImage(named: "logo")) // 이미지 이름을 넣어주세요
-        view.addSubview(logo)
-        
-        // Label
-        let label = UILabel()
-        label.text = "공부하개미"
-        label.font = UIFont(name: "Pretendard-Medium", size: 30) // 폰트 및 사이즈 설정
-        label.textColor = UIColor(named: "launchScreenFontColor")
-        label.textAlignment = .center
-        view.addSubview(label)
-        
-        let company = UILabel()
-        company.text = "Ⓒ 8tTruck"
-        company.font = UIFont(name: "Pretendard-Medium", size: 16)
-        company.textColor = UIColor(named: "launchScreenFontColor")
-        company.textAlignment = .center
-        view.addSubview(company)
-        
-        // SnapKit을 사용하여 Auto Layout 설정
-        logo.snp.makeConstraints { make in
-            make.centerX.equalToSuperview() // 화면 가운데에 위치
-            make.centerY.equalToSuperview().offset(-50) // 세로 중앙에서 위로 50 포인트 이동
-            make.width.equalTo(view.snp.width).multipliedBy(0.5) // 화면 너비의 50%
-            make.height.equalTo(logo.snp.width).multipliedBy(1.0) // 가로세로 비율 1:1로 유지
-        }
-        
-        label.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(logo.snp.bottom).offset(10) // 로고 아래에서 20 포인트 떨어진 곳에 위치
-        }
-        
-        company.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalToSuperview().offset(-40)
-        }
-    }
-}
-
 extension SceneDelegate {
     
     // MARK: - 특정 ViewController로 이동하는 메소드
@@ -205,6 +113,55 @@ extension SceneDelegate {
             }
         default:
             return LoginViewController()
+        }
+    }
+}
+
+extension SceneDelegate {
+    
+    func checkLogin() {
+        print("로그인 유저인지 확인")
+        if let user = Auth.auth().currentUser {
+            print("currentUser임")
+            AuthenticationManager.shared.checkEmailVerifiedForLogin { isEmailVerified in
+                if isEmailVerified {
+                    self.isLogined = true
+                    print("애플 또는 이메일 자동 로그인 성공: \(user.email ?? "")")
+                    self.navigateToMainScreen()
+                } else {
+                    print("자동 로그인 실패. else 시작")
+                    FirestoreManager.shared.getLoginMethod { loginMethod in
+                        if loginMethod == "kakao" {
+                            print("카카오 계정임")
+                            self.isLogined = true
+                            self.navigateToMainScreen()
+                        } else {
+                            print("카카오 계정 이외의 알 수 없는 무언가")
+                            self.navigateToLoginScreen()
+                        }
+                    }
+                }
+            }
+        } else {
+            print("최종 else 진입")
+            self.navigateToLoginScreen()
+        }
+    }
+    
+    
+    func navigateToMainScreen() {
+        let bottomTabBarVC = BottomTabBarViewController()
+        if let window = self.window {
+            window.rootViewController = bottomTabBarVC
+            window.makeKeyAndVisible()
+        }
+    }
+    
+    func navigateToLoginScreen() {
+        let bottomTabBarVC = LoginViewController()
+        if let window = self.window {
+            window.rootViewController = bottomTabBarVC
+            window.makeKeyAndVisible()
         }
     }
 }
